+++ HISTORY.md
@@ -2,6 +2,7 @@ unreleased
 ==========
 
 - new: nocache now sets the Expires and Pragma headers
+- new: nocache now allows you to crush ETags
 
 0.3.2 / 2014-06-30
 ==================